Добавил дополнительные проверки в ProfessionService и DepartmentService при удалении, добавил исключения CheckConstraintsException и CanNotDeleteException, добавил тест на удаление професии.

+++ src/main/java/org/tonkushin/example/service/DepartmentServiceImpl.java
@@ -1,9 +1,12 @@
 package org.tonkushin.example.service;
 
 import org.springframework.stereotype.Service;
+import org.tonkushin.example.exception.CanNotDeleteException;
+import org.tonkushin.example.exception.CheckConstraintsException;
 import org.tonkushin.example.exception.ItemNotFoundException;
 import org.tonkushin.example.model.Department;
 import org.tonkushin.example.repository.DepartmentRepository;
+import org.tonkushin.example.repository.PersonRepository;
 
 import java.util.List;
 
@@ -15,9 +18,11 @@ import java.util.List;
 public class DepartmentServiceImpl implements DepartmentService {
 
     private final DepartmentRepository repository;
+    private final PersonRepository personRepository;
 
-    public DepartmentServiceImpl(DepartmentRepository repository) {
+    public DepartmentServiceImpl(DepartmentRepository repository, PersonRepository personRepository) {
         this.repository = repository;
+        this.personRepository = personRepository;
     }
 
     @Override
@@ -39,11 +44,16 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public void delete(Department item) {
-        repository.delete(item);
+        delete(item.getId());
     }
 
     @Override
     public void delete(Long id) {
+        // Проверяем, что нет сотрудников, относящихся к указанному отделу
+        if (personRepository.countAllByDepartment_Id(id) > 0) {
+            throw new CanNotDeleteException();
+        }
+
         repository.deleteById(id);
     }
 
@@ -56,14 +66,14 @@ public class DepartmentServiceImpl implements DepartmentService {
     private void checkConstraints(Department item) {
         // Проверяем ограничения по полю Наименование
         if (item.getName() == null || item.getName().isEmpty()) {
-            throw new RuntimeException("Поле 'Наименование' не должно быть пустым");
+            throw new CheckConstraintsException("Поле 'Наименование' не должно быть пустым");
         } else if (item.getName() != null && item.getName().length() > 255) {
-            throw new RuntimeException("Поле 'Наименование' не должно превышать 255 символов");
+            throw new CheckConstraintsException("Поле 'Наименование' не должно превышать 255 символов");
         }
 
         // Проверяем ограничения по полю Примечание
         if (item.getNotes() != null && item.getNotes().length() > 255) {
-            throw new RuntimeException("Поле 'Примечание' не должно превышать 255 символов");
+            throw new CheckConstraintsException("Поле 'Примечание' не должно превышать 255 символов");
         }
     }
 }
