+++ src/main/java/org/tonkushin/example/service/ProfessionServiceImpl.java
@@ -1,8 +1,11 @@
 package org.tonkushin.example.service;
 
 import org.springframework.stereotype.Service;
+import org.tonkushin.example.exception.CanNotDeleteException;
+import org.tonkushin.example.exception.CheckConstraintsException;
 import org.tonkushin.example.exception.ItemNotFoundException;
 import org.tonkushin.example.model.Profession;
+import org.tonkushin.example.repository.PersonRepository;
 import org.tonkushin.example.repository.ProfessionRepository;
 
 import java.util.List;
@@ -16,9 +19,11 @@ import java.util.Optional;
 public class ProfessionServiceImpl implements ProfessionService {
 
     private final ProfessionRepository repository;
+    private final PersonRepository personRepository;
 
-    public ProfessionServiceImpl(ProfessionRepository repository) {
+    public ProfessionServiceImpl(ProfessionRepository repository, PersonRepository personRepository) {
         this.repository = repository;
+        this.personRepository = personRepository;
     }
 
     @Override
@@ -40,11 +45,15 @@ public class ProfessionServiceImpl implements ProfessionService {
 
     @Override
     public void delete(Profession item) {
-        repository.delete(item);
+        delete(item.getId());
     }
 
     @Override
     public void delete(Long id) {
+        // Проверяем, что нет сотрудников с этой профессией
+        if (personRepository.countAllByProfession_Id(id) > 0) {
+            throw new CanNotDeleteException();
+        }
         repository.deleteById(id);
     }
 
@@ -57,21 +66,21 @@ public class ProfessionServiceImpl implements ProfessionService {
     private void checkConstraints(Profession item) {
         // Проверяем ограничения по полю Наименование
         if (item.getName() == null || item.getName().isEmpty()) {
-            throw new RuntimeException("Поле 'Наименование' не должно быть пустым");
+            throw new CheckConstraintsException("Поле 'Наименование' не должно быть пустым");
         } else if (item.getName() != null && item.getName().length() > 255) {
-            throw new RuntimeException("Поле 'Наименование' не должно превышать 255 символов");
+            throw new CheckConstraintsException("Поле 'Наименование' не должно превышать 255 символов");
         }
 
         // Проверяем, что профессии с таким именем не существует
         Optional<Profession> opt = repository.findByName(item.getName());
 
         if (opt.isPresent() && opt.get().getId() != item.getId()) {
-            throw new RuntimeException(String.format("Профессия с именем '%s' уже существует", item.getName()));
+            throw new CheckConstraintsException(String.format("Профессия с именем '%s' уже существует", item.getName()));
         }
 
         // Проверяем ограничения по полю Примечание
         if (item.getNotes() != null && item.getNotes().length() > 255) {
-            throw new RuntimeException("Поле 'Примечание' не должно превышать 255 символов");
+            throw new CheckConstraintsException("Поле 'Примечание' не должно превышать 255 символов");
         }
     }
 }
